Добавлена сборка зависимостей
