+++ src/game.js
@@ -1,6 +1,6 @@
 'use strict';
 
-window.Game = (function() {
+define(function() {
   /**
    * @const
    * @type {number}
@@ -751,4 +751,4 @@ window.Game = (function() {
   Game.Verdict = Verdict;
 
   return Game;
-})();
+});
