+++ src/main.js
@@ -1,13 +1,9 @@
 'use strict';
 
-require('./form');
-require('./game');
-require('./reviews.js');
-
-(function() {
-  var game = new window.Game(document.querySelector('.demo'));
+define(['./form', './game', './reviews'], function(form, Game) {
+  var game = new Game(document.querySelector('.demo'));
   game.initializeLevelAndStart();
-  game.setGameStatus(window.Game.Verdict.INTRO);
+  game.setGameStatus(Game.Verdict.INTRO);
 
   var formOpenButton = document.querySelector('.reviews-controls-new');
 
@@ -15,13 +11,13 @@ require('./reviews.js');
   formOpenButton.onclick = function(evt) {
     evt.preventDefault();
 
-    window.form.open(function() {
-      game.setGameStatus(window.Game.Verdict.PAUSE);
+    form.open(function() {
+      game.setGameStatus(Game.Verdict.PAUSE);
       game.setDeactivated(true);
     });
   };
 
-  window.form.onClose = function() {
+  form.onClose = function() {
     game.setDeactivated(false);
   };
-})();
+});
